Move TestMethodInvoker to the common test module
